Refactor vocabulary feature: Move components and API hooks to a new directory structure, update imports, and remove unused files.
