+++ features/vocabulary/components/card-vocabulary.tsx
